get all the requests that the useposted - by token

+++ routes/users.js
@@ -4,11 +4,6 @@ const { auth, authAdmin } = require("../middlewares/auth");
 const { UserModel, validUser, validLogin, createToken } = require("../models/userModel")
 const router = express.Router();
 
-// dont know why we need it not used
-router.get("/", async (req, res) => {
-  res.status(201).json({ msg: "Users work" })
-})
-
 // get current user details
 router.get("/myInfo", auth, async (req, res) => {
   try {
@@ -87,29 +82,71 @@ router.put("/changePassword", auth, async (req, res) => {
   // if not throw
   // if correct hash the new password and save it
 })
+router.put("/forgotPassword", auth, async (req, res) => {
+  // not sure if good
+  // only the user itself can change his password
+  // need to validate the token
+  // get a password and make sure the password is correct
+  // if not throw
+  // if correct hash the new password and save it
+})
+
+
+router.put("/:idEdit", auth, async(req,res) =>{
+  let validBody = validUser(req.body);
+  if (validBody.error) {
+    return res.status(400).json(validBody.error.details);
+  }
+  try {
+    let idEdit = req.params.idEdit;
+    let data;
+    if (req.tokenData.role === "admin") {
+      data = await UserModel.updateOne({ _id: idEdit }, req.body)
+    }
+    else if (idEdit === req.tokenData._id) {
+      data = await UserModel.updateOne({ _id: idEdit }, req.body)
+    }
+    if (!data) {
+      return res.status(400).json({ err: "This operation is not enabled !" })
+    }
+    let user = await UserModel.findOne({ _id: idEdit });
+    user.password = await bcrypt.hash(user.password, 10);
+    await user.save()
+    res.status(200).json({ msg: data })
+  }
+  catch (err) {
+    console.log(err);
+    res.status(400).json({ err })
+  }
+})
+
+// delete user accont - by user token
+router.delete("/deleteAccount", auth, async (req, res) => {
+  try {
+    if(req.tokenData.role === "admin"){
+      return res.status(400).json("can't delete an admin user");
+    }
+    else{
+      let data = await UserModel.deleteOne({ _id: req.tokenData._id })
+      res.status(201).json(data);
+    }
+  }
+  catch (err) {
+    console.log(err);
+    res.status(500).json({ msg: "err to delete your account", err })
+  }
+})
 
 // delete spesific user by admin token
 router.delete("/:idDel", authAdmin, async (req, res) => {
   try {
-    let delId = req.params.delId;
+    let delId = req.params.idDel;
     let data = await UserModel.deleteOne({ _id: delId });
     res.status(201).json(data);
   }
   catch (err) {
     console.log(err);
     res.status(500).json({ msg: "err to delete this user by id", err })
-  }
-})
-
-// delete user accont - by user token
-router.delete("/deleteAccount", auth, async (req, res) => {
-  try {
-    let data = await UserModel.deleteOne({ _id: req.tokenData._id })
-    res.status(201).json(data);
-  }
-  catch (err) {
-    console.log(err);
-    res.status(500).json({ msg: "err to delete your account", err })
   }
 })
